requirements.txt
fastapi
uvicorn
python-dotenv
openai==1.3.0
pydantic
pdfplumber
pytesseract
Pillow
pandas

